+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [1.3.0](https://github.com/kyndryl-design-system/shidoka-charts/compare/v1.2.2...v1.3.0) (2024-04-11)
+
+
+### Features
+
+* **doughnut:** make doughnut center labels customizable ([#19](https://github.com/kyndryl-design-system/shidoka-charts/issues/19)) ([8ec3b4c](https://github.com/kyndryl-design-system/shidoka-charts/commit/8ec3b4c853dfb1b6d437f76420b02656d852a03e))
+
+## [1.2.2](https://github.com/kyndryl-design-system/shidoka-charts/compare/v1.2.1...v1.2.2) (2024-04-10)
+
+
+### Bug Fixes
+
+* **chart:** prevent re-render on init, remove certain plugins for third party chart types ([81d9400](https://github.com/kyndryl-design-system/shidoka-charts/commit/81d94001788a79e80cd0d2c28ff5d4170eeaad6d))
+
 ## [1.2.1](https://github.com/kyndryl-design-system/shidoka-charts/compare/v1.2.0...v1.2.1) (2024-04-08)